[Fix] 약물 조회 오류 수정 & 보호관계 생성 시 결제유무 에러코드 추가

+++ src/main/java/com/cesco/pillintime/api/relation/service/RelationService.java
@@ -39,6 +39,10 @@ public class RelationService {
 
         Member targetMember = request.getSender();
 
+        if (!targetMember.isSubscriber()) {
+            throw new CustomException(ErrorCode.MANAGER_IS_NOT_SUBSCRIBER);
+        }
+
         Relation relation = new Relation(targetMember, requestMember);
         relationRepository.save(relation);
         requestRepository.delete(request);
